Bombs are now exploited at exact position-updated

+++ core/src/de/tum/cit/ase/bomberquest/Actors/Player.java
@@ -146,15 +146,16 @@ public class Player implements Drawable {
     }
 
     @Override
-    public int getX() {
+    public float getX() {
         // The x-coordinate of the player is the x-coordinate of the hitbox (this can change every frame).
-        return (int) hitbox.getPosition().x;
+        return hitbox.getPosition().x;
     }
 
     @Override
-    public int getY() {
+    public float
+    getY() {
         // The y-coordinate of the player is the y-coordinate of the hitbox (this can change every frame).
-        return (int) hitbox.getPosition().y;
+        return hitbox.getPosition().y;
     }
 
     private enum Direction {
