+++ core/src/de/tum/cit/ase/bomberquest/map/Exit.java
@@ -8,6 +8,7 @@ import com.badlogic.gdx.physics.box2d.World;
 import de.tum.cit.ase.bomberquest.Actors.Santa;
 import de.tum.cit.ase.bomberquest.texture.Drawable;
 import de.tum.cit.ase.bomberquest.texture.Textures;
+import de.tum.cit.ase.bomberquest.map.GameMap;
 
 public class Exit implements Drawable {
 
@@ -83,13 +84,13 @@ public class Exit implements Drawable {
     }
 
     @Override
-    public int getX() {
-        return (int) x;
+    public float getX() {
+        return x;
     }
 
     @Override
-    public int getY() {
-        return (int) y;
+    public float getY() {
+        return y;
     }
 
     public boolean ExitOpened() {
